+++ src/textual/layouts/stream.py
@@ -37,7 +37,7 @@ class StreamLayout(Layout):
         parent.pre_layout(self)
         if not children:
             return []
-        viewport = parent.app.size
+        viewport = parent.app.viewport_size
 
         _Region = Region
         _WidgetPlacement = WidgetPlacement
@@ -81,3 +81,37 @@ class StreamLayout(Layout):
             y += height
 
         return placements
+
+    def get_content_width(self, widget: Widget, container: Size, viewport: Size) -> int:
+        """Get the optimal content width by arranging children.
+
+        Args:
+            widget: The container widget.
+            container: The container size.
+            viewport: The viewport size.
+
+        Returns:
+            Width of the content.
+        """
+        return widget.scrollable_content_region.width
+
+    def get_content_height(
+        self, widget: Widget, container: Size, viewport: Size, width: int
+    ) -> int:
+        """Get the content height.
+
+        Args:
+            widget: The container widget.
+            container: The container size.
+            viewport: The viewport.
+            width: The content width.
+
+        Returns:
+            Content height (in lines).
+        """
+        if widget._nodes:
+            arrangement = widget._arrange(Size(width, 0))
+            height = arrangement.total_region.height
+        else:
+            height = 0
+        return height