+++ src/textual/layouts/vertical.py
@@ -23,7 +23,7 @@ class VerticalLayout(Layout):
         parent.pre_layout(self)
         placements: list[WidgetPlacement] = []
         add_placement = placements.append
-        viewport = parent.app.size
+        viewport = parent.app.viewport_size
 
         child_styles = [child.styles for child in children]
         box_margins: list[Spacing] = [